# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart Tiffin & Student Analytics Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students/profile:
    get:
      operationId: getStudentProfile
      tags: [students]
      summary: Get student profile
      responses:
        "200":
          description: Student profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
    put:
      operationId: updateStudentProfile
      tags: [students]
      summary: Update student profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentProfileInput"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
...
components:
  schemas:
...
    StudentProfile:
      type: object
      required: [id, userId, name, email]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        college:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        profilePicture:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    StudentProfileInput:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        college:
          type: string
        address:
          type: string
...